Remove unnecessary casts

+++ src/main/java/grondag/canvas/terrain/BuiltRenderRegion.java
@@ -318,11 +318,11 @@ public class BuiltRenderRegion {
 
 				if (Configurator.batchedChunkRender) {
 					collector.sortQuads(
-						(float) cameraPos.x - TerrainModelSpace.renderCubeOrigin(origin.getX()),
-						(float) cameraPos.y - TerrainModelSpace.renderCubeOrigin(origin.getY()),
-						(float) cameraPos.z - TerrainModelSpace.renderCubeOrigin(origin.getZ()));
+						cameraPos.x - TerrainModelSpace.renderCubeOrigin(origin.getX()),
+						cameraPos.y - TerrainModelSpace.renderCubeOrigin(origin.getY()),
+						cameraPos.z - TerrainModelSpace.renderCubeOrigin(origin.getZ()));
 				} else {
-					collector.sortQuads((float) cameraPos.x - origin.getX(), (float) cameraPos.y - origin.getY(), (float) cameraPos.z - origin.getZ());
+					collector.sortQuads(cameraPos.x - origin.getX(), cameraPos.y - origin.getY(), cameraPos.z - origin.getZ());
 				}
 
 				regionData.translucentState = collector.saveState(state);
